+++ dsr_agents/test/test_tf_agent_integration.cpp
@@ -33,11 +33,11 @@ TEST_F(DsrUtilTest, tfAgentIntegration) {
   auto pub_thread = std::thread([&]() {rclcpp::spin(pub_node->get_node_base_interface());});
 
   // Create and configure the tf_agent node
-  auto agent_node = std::make_shared<dsr_agents::TFAgent>();
+  auto node_agent = std::make_shared<dsr_agents::TFAgent>();
   // Set some parameters
-  agent_node->declare_parameter("dsr_input_file", rclcpp::ParameterValue(test_file_));
-  agent_node->configure();
-  agent_node->activate();
+  node_agent->declare_parameter("dsr_input_file", rclcpp::ParameterValue(test_file_));
+  node_agent->configure();
+  node_agent->activate();
 
   // Create a tf message
   tf2_msgs::msg::TFMessage tf_msg;
@@ -50,14 +50,14 @@ TEST_F(DsrUtilTest, tfAgentIntegration) {
   tf_pub->publish(tf_msg);
 
   // Spin the tf_agent node
-  rclcpp::spin_some(agent_node->get_node_base_interface());
+  rclcpp::spin_some(node_agent->get_node_base_interface());
 
   // Check the results: now, the tf should have a subscription
   EXPECT_EQ(tf_pub->get_subscription_count(), 1);
 
   // Deactivate the nodes
-  agent_node->deactivate();
-  agent_node->cleanup();
+  node_agent->deactivate();
+  node_agent->cleanup();
   pub_node->deactivate();
   rclcpp::shutdown();
   // Have to join thread after rclcpp is shut down otherwise test hangs.