+++ dsr_util/CHANGELOG.rst
@@ -7,7 +7,7 @@ Changelog for package dsr_util
 * Update license to Apache 2.0.
 * Improve formating.
 * CMakeLists.txt use modern idioms.
-* Move DSR callback functions to agent_node.
+* Move DSR callback functions to node_agent.
 * Move action_agent to dsr_util.
 * Converted agent class to lifecycle node.
 * Added unit tests.
@@ -20,7 +20,7 @@ Changelog for package dsr_util
 * Initial release.
 * Create README.md.
 * Added qt_executor class (.hpp and .cpp files).
-* Added agent_node class (.hpp and .cpp files).
+* Added node_agent class (.hpp and .cpp files).
 * Added ros_to_dsr_types.hpp.
 * Added worlds.
 * Contributors: Alberto Tudela